oled/pyproject.toml
[project]
name = "renderer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "psutil>=7.1.0",
    "pyserial>=3.5",
    "pyte>=0.8.2",
]
